Refactor DojoUtil into separate module
