+++ libs/quickdetect/__init__.py
@@ -1,0 +1,1 @@
+from .DojoUtil import DojoUtil
